+++ src/asciidoc/templates/Example.tsx
@@ -1,43 +1,40 @@
-import type { Block } from '@asciidoctor/core'
 import cn from 'classnames'
 import parse from 'html-react-parser'
 
 import { Content } from '../'
-import { CaptionedTitle } from './util'
-import { getLineNumber, getRole } from './util'
+import { type Block, isOption } from '../utils/prepareDocument'
+import { Title } from './util'
 
 const Example = ({ node }: { node: Block }) => {
-  if (node.isOption('collapsible')) {
-    const isCollapsible = node.isOption('collapsible')
-    const title = node.getTitle() || 'Details'
-    const isOpen = node.isOption('open') ? true : undefined
+  const isCollapsible = isOption(node.attributes, 'collapsible')
+  if (isCollapsible) {
+    const title = node.title || 'Details'
+    const isOpen = node.attributes.open ? true : undefined
 
-    if (isCollapsible) {
-      return (
-        <details
-          className={getRole(node)}
-          open={isOpen}
-          {...getLineNumber(node)}
-          {...(node.getId() ? { id: node.getId() } : {})}
-        >
-          <summary className="title">{parse(title)}</summary>
-          <div className="content">
-            <Content blocks={node.getBlocks()} />
-          </div>
-        </details>
-      )
-    }
+    return (
+      <details
+        className={node.role}
+        open={isOpen}
+        {...(node.lineNumber ? { 'data-lineno': node.lineNumber } : {})}
+        {...(node.id ? { id: node.id } : {})}
+      >
+        <summary className="title">{parse(title)}</summary>
+        <div className="content">
+          <Content blocks={node.blocks} />
+        </div>
+      </details>
+    )
   }
 
   return (
     <div
-      className={cn('exampleblock', getRole(node))}
-      {...getLineNumber(node)}
-      {...(node.getId() ? { id: node.getId() } : {})}
+      className={cn('exampleblock', node.role)}
+      {...(node.lineNumber ? { 'data-lineno': node.lineNumber } : {})}
+      {...(node.id ? { id: node.id } : {})}
     >
-      <CaptionedTitle node={node} />
+      <Title text={node.title} />
       <div className="content">
-        <Content blocks={node.getBlocks()} />
+        <Content blocks={node.blocks} />
       </div>
     </div>
   )