+++ src/asciidoc/templates/Literal.tsx
@@ -1,16 +1,21 @@
-import type { Block } from '@asciidoctor/core'
+import { useConverterContext } from '..'
+import { type LiteralBlock } from '../utils/prepareDocument'
+import { Title } from './util'
 
-import { Title, getLineNumber } from './util'
+const Literal = ({ node }: { node: LiteralBlock }) => {
+  const { document } = useConverterContext()
 
-const Literal = ({ node }: { node: Block }) => {
-  const docAttrs = node.getDocument().getAttributes()
-  const nowrap = docAttrs['prewrap'] === undefined || node.isOption('nowrap')
+  const docAttrs = document.attributes || {}
+  const nowrap = node.attributes.nowrap || docAttrs['prewrap'] === undefined
 
   return (
-    <div className="literalblock" {...getLineNumber(node)}>
-      <Title node={node} />
+    <div
+      className="literalblock"
+      {...(node.lineNumber ? { 'data-lineno': node.lineNumber } : {})}
+    >
+      <Title text={node.title} />
       <div className="content">
-        <pre className={nowrap ? 'nowrap' : ''}>{node.getSource()}</pre>
+        <pre className={nowrap ? 'nowrap' : ''}>{node.source}</pre>
       </div>
     </div>
   )