+++ src/asciidoc/templates/util.tsx
@@ -1,16 +1,4 @@
-import type { AbstractBlock, Block, List } from '@asciidoctor/core'
 import parse from 'html-react-parser'
 
-export const Title = ({ node }: { node: AbstractBlock | Block | List }) =>
-  node.hasTitle() ? <div className="title">{parse(node.getTitle() || '')}</div> : null
-
-export const CaptionedTitle = ({ node }: { node: Block | Block | List }) =>
-  node.hasTitle() ? (
-    <div className="title">{parse(node.getCaptionedTitle() || '')}</div>
-  ) : null
-
-export const getRole = (node: AbstractBlock | Block | List) =>
-  typeof node.getRole() === 'string' ? node.getRole() : undefined
-
-export const getLineNumber = (node: AbstractBlock | Block | List) =>
-  node.getLineNumber() ? { 'data-lineno': node.getLineNumber() } : {}
+export const Title = ({ text }: { text: string | undefined }) =>
+  text ? <div className="title">{parse(text)}</div> : null